Cambio de tipo de variable, correción de errores y creación de la funcionalidad de agregar personas

+++ app/src/main/java/com/aplicada2/tareai/data/database/entities/Persona.kt
@@ -8,5 +8,5 @@ data class Persona (
     @PrimaryKey(autoGenerate = true)
     val PersonaId: Int,
     val Nombres: String,
-    val Balance: Float
+    val Balance: Double
 )
